requirements.txt
Django>=4.2,<5.0
python-dotenv
#mozilla-django-oidc
PyJWT
debugpy>=1.8.0
boto3
django-crispy-forms
crispy-bootstrap5
django-widget-tweaks>=1.4.8
PyPDF2>=3.0.0
reportlab>=3.6.0
whitenoise
django-auth-adfs
gunicorn
psycopg[binary]